Add query_builder.py and add fixed filtering and searching methods

+++ src/crud/crud_movies.py
@@ -1,14 +1,12 @@
-from typing import Any
-
-from fastapi import HTTPException, Query
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
-from src.models.models_movies import Movie, Review
-from src.schemas import movies as schemas
+from src import schemas, models
+from src.query_builder import get_filtered_query, search_query
 
 
-def get_movie_by_id(movie_id: int, db: Session) -> Movie:
-    db_movie = get_filtered_query(Movie, db.query(Movie), {Movie.id.key: movie_id}).first()
+def get_movie_by_id(movie_id: int, db: Session) -> schemas.MovieBase:
+    db_movie = get_filtered_query(models.Movie, db.query(models.Movie), {models.Movie.id.key: movie_id}).first()
     if not db_movie:
         raise HTTPException(status_code=400, detail=f"Movie ID {movie_id} not found")
     return db_movie
@@ -16,33 +14,27 @@ def get_movie_by_id(movie_id: int, db: Session) -> Movie:
 
 def get_movies(args: schemas.MoviesGetRequest, db: Session) -> list[schemas.MovieBase]:
     filter_fields = {
-        Movie.title.key: args.title,
+        models.Movie.watched.key: args.watched,
     }
+    filtered_movies = get_filtered_query(models.Movie, db.query(models.Movie), filter_fields)
+    searched_movies = search_query(
+        filtered_movies,
+        models.Movie,
+        args.search,
+        [models.Movie.title.key, models.Movie.id.key],
+    )
 
-    movie_objects = get_filtered_query(Movie, db.query(Movie), filter_fields).all()
+    movie_objects = searched_movies.all()
     return [schemas.MovieBase(**item.__dict__) for item in movie_objects]
 
 
-def get_filtered_query(table, query: Query, filter_fields: dict[str, Any]) -> Query:
-    if not filter_fields:
-        return query
-
-    for attr, value in filter_fields.items():
-        filter_obj = getattr(table, attr, None)
-        if filter_obj and value:
-            query = query.filter(filter_obj == value)
-
-    return query
-
-
-def create_movie(movie: schemas.MovieCreate, db: Session):
-    movie_title = movie.title.lower()
-    movie_exists = db.query(db.query(Movie).filter(Movie.title == movie_title).exists()).scalar()
+def create_movie(movie: schemas.MovieCreate, db: Session) -> models.Movie:
+    movie_exists = db.query(db.query(models.Movie).filter(models.Movie.title == movie.title).exists()).scalar()
     if movie_exists:
-        raise HTTPException(status_code=400, detail=f"Movie with title - {movie_title} already created")
+        raise HTTPException(status_code=400, detail=f"Movie with title - {movie.title} already created")
 
-    db_movie = Movie(
-        title=movie_title,
+    db_movie = models.Movie(
+        title=movie.title,
         description=movie.description,
         watched=movie.watched,
         rating=movie.rating)
@@ -54,6 +46,8 @@ def create_movie(movie: schemas.MovieCreate, db: Session):
 
 def update_movie_by_id(movie_id: int, movie: schemas.MovieUpdate, db: Session):
     db_movie = get_movie_by_id(movie_id, db)
+    if not db_movie:
+        raise HTTPException(status_code=400, detail=f"Movie with ID {movie_id} does not exist")
 
     for key, value in schemas.MovieUpdate(**movie.__dict__):
         setattr(db_movie, key, value)
@@ -65,5 +59,7 @@ def update_movie_by_id(movie_id: int, movie: schemas.MovieUpdate, db: Session):
 
 def delete_movie_by_id(movie_id: int, db: Session) -> None:
     db_movie = get_movie_by_id(movie_id, db)
+    if not db_movie:
+        raise HTTPException(status_code=400, detail=f"Movie with ID {movie_id} does not exist")
     db.delete(db_movie)
     db.commit()
