+++ src/crud/crud_reviews.py
@@ -1,26 +1,12 @@
-from typing import Any
-
-from fastapi import HTTPException, Query
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
-from src.models.models_movies import Review
-from src.schemas import reviews as schemas
+from src import models, schemas
+from src.query_builder import get_filtered_query, search_query
 
 
-def get_filtered_query(table, query: Query, filter_fields: dict[str, Any]) -> Query:
-    if not filter_fields:
-        return query
-
-    for attr, value in filter_fields.items():
-        filter_obj = getattr(table, attr, None)
-        if filter_obj and value:
-            query = query.filter(filter_obj == value)
-
-    return query
-
-
-def get_review_by_id(review_id: int, db: Session) -> Review:
-    db_review = get_filtered_query(Review, db.query(Review), {Review.id.key: review_id}).first()
+def get_review_by_id(review_id: int, db: Session) -> models.Review:
+    db_review = get_filtered_query(models.Review, db.query(models.Review), {models.Review.id.key: review_id}).first()
     if not db_review:
         raise HTTPException(status_code=400, detail=f"Movie ID {review_id} not found")
     return db_review
@@ -28,22 +14,28 @@ def get_review_by_id(review_id: int, db: Session) -> Review:
 
 def get_reviews(args: schemas.ReviewsGetRequest, db: Session) -> list[schemas.ReviewBase]:
     filter_fields = {
-        Review.title.key: args.search,
+        models.Review.movie_id.key: args.movie_id,
     }
 
-    review_objects = get_filtered_query(Review, db.query(Review), filter_fields).all()
-    return [schemas.ReviewBase(**item.__dict__) for item in review_objects]
-
-
-def get_reviews_by_movie(movie_id: int, db: Session) -> list[schemas.ReviewBase]:
-    reviews = db.query(Review).filter(Review.movie_id == movie_id).all()
-    return [schemas.ReviewBase(**review.__dict__) for review in reviews]
+    filtered_reviews = get_filtered_query(models.Review, db.query(models.Review), filter_fields)
+    searched_reviews = search_query(
+        filtered_reviews,
+        models.Review,
+        args.search,
+        [models.Review.text.key, models.Review.movie_id.key, models.Review.id.key, models.Review.title.key]
+    )
+    return [schemas.ReviewBase(**item.__dict__) for item in searched_reviews]
 
 
 def create_review(review: schemas.ReviewCreate, movie_id: int, db: Session):
-    db_review = Review(
+    movie_id_exists = db.query(db.query(models.Movie).filter(models.Movie.id == movie_id).exists()).scalar()
+    if not movie_id_exists:
+        raise HTTPException(status_code=400, detail=f"Movie ID {movie_id} does not exist")
+    db_review = models.Review(
+        title=review.title,
         text=review.text,
         movie_id=movie_id,
+
     )
     db.add(db_review)
     db.commit()
@@ -53,9 +45,12 @@ def create_review(review: schemas.ReviewCreate, movie_id: int, db: Session):
 
 def update_review_by_id(review_id: int, review: schemas.ReviewUpdate, db: Session):
     db_review = get_review_by_id(review_id, db)
+    if not db_review:
+        raise HTTPException(status_code=400, detail=f"Review with ID {review_id} does not exist")
 
     for key, value in schemas.ReviewUpdate(**review.__dict__):
         setattr(db_review, key, value)
+
     db.add(db_review)
     db.commit()
     db.refresh(db_review)
@@ -64,5 +59,7 @@ def update_review_by_id(review_id: int, review: schemas.ReviewUpdate, db: Sessio
 
 def delete_review_by_id(review_id: int, db: Session) -> None:
     db_review = get_review_by_id(review_id, db)
+    if not db_review:
+        raise HTTPException(status_code=400, detail=f"Review with ID {review_id} does not exist")
     db.delete(db_review)
     db.commit()
